development
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@86 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/gameevent.h
@@ -55,7 +55,7 @@ private:
     Artifact::artifact_t artifact;
     bool computer;
     u16 first;
-    u8 subsequent;
+    u16 subsequent;
     u8 colors;
     std::string message;
 };
